[FIX] Leaking old `syncthing-server` instance, breaking multi-client
Without cleanup an old instance was picked up. For a scenario when a user
enters wrong data a broken client was created and put to `syncthing--servers`
global variable. That one wasn't cleaned before and using

`(car (or syncthing--servers ...))`

always picked such instance even though a new one, let's say with a correct
token, had already been created.

For multi-client behavior such as combining `syncthing` and
`syncthing-with-base` when creating a demo view the same URL (of the first
client put to `syncthing--servers`) was picked up.

Ref:
* https://github.com/KeyWeeUsr/emacs-syncthing/issues/2#issuecomment-1866530545