[REF] Use `<tab>` for `widget-forward`
This is adjustable with `syncthing-old-tab-behavior` although
`<enter>`/`<space>` seems like a better way to handle it even though it moves
away from how Org mode does it commonly. `<backtab>` is also changed. The
original functionality is moved under Ctrl, so that `widget-keymap` is used for
the navigation between widgets and text hints.