[REF] Don't constantly reset the current buffer
For some reason, when I try to refresh the buffer from a separate
thread, the deleted code causes newlines between widgets to disappear if
and only if the syncthing buffer isn't focused. When I remove this code,
everything works correctly.

Furthermore, it appears that none of this code is actually necessary.
Nor does it do what it appears to do because, e.g., `syncthing-buffer'
will be nil unless the current buffer IS the syncthing buffer.